Add test of the complement conjecture for 1D and 2D arrays.
